Set time_zone to UTC on MySQL
MySQL uses UTC for all date fields but other operations like `now()` use
the timezone of the server. This patch changes the timezone of the
connection to UTC when the connection is first established.

There's space for improvement here. For instance, we could try to avoid
doing this for every new connection and let the operator configure it on
the server side. We could also use `CONVERT_TZ` on every query using
`now()`.

Change-Id: I2e8415dbb9a6ce19d3a66f3226756094b3e0f6a4
Partial-bug: #1288820