Remove MySQL-python from the dependency list
MySQL is not a recommended database for production environments. The
dependency to the mysql python library was added to be able to run
Marconi on the gate.

Since we're not willing to promote sqlalchemy as the recommended driver,
it makes sense to remove the mysql dependency from the dependencies
list.

Closes-bug: #1308500

Change-Id: Ic83707cb145d752b30e285f1a25376525a80079c