fix(metadata): Fix inconsistent error response under v1.1
This patch modifies the v1.1 `on_put` responder to raise an
instance of falcon.HTTPError, including setting the Allow
header as required by RFC 2616.

This ensures that the structured JSON response body is
consistent with error bodies returned elsewhere in the
API.

A test was also added to cover the v1.1 implementation
of `on_put`.

Change-Id: I05e15cd3c676db8f9fc3fb7d1ba6fcd31613fc30
Partially-Implements: blueprint api-v1.1-remove-queue-metadata