Make storage/pooling reflect storage/base
Controllers wrappers for pooled context wasn't fully compatible with the
APIs defined in storage/base. Order of parameters was different and
default values was not specified or specified incorrectly.  In
particular in pooled setup default limits were None in message and claim
controllers.  This commit changes them to be the same as in non-pooled
setup.

Also tests for testing default limits were added.

Change-Id: I2b83383975986179813df7a897d78852a532564e
Closes-bug: #1324587