Mark sqlalchemy's data driver as deprecated
In the Kilo release, we'll be removing the *data* driver for sqlalchemy.
The motivations behind this are that we don't consider this driver good
for production nor the team has the bandwidth to maintain it.

In order to allow users to migrate to another driver, we're marking this
driver as deprecated. In the kilo release, it'll be pulled out into its
own repository and left there for other folks to maintain it if they
consider it useful.

Change-Id: I4efb9ec46080f111705b92e2395b88c64c66e1f8