Merge metadata back into the queue endpoint
During Icehouse, we decided to pull the metadata out of the queue
endpoint by adding a new endpoint `/metadata`. This worked well because
we didn't have a clear view of what the scope of metadata is. Now that
we want to use it and make it a key piece of the queues, we've found
that the workflow is not as clear as we'd like. The reason being that
some key items of the metadata - flavor, for example - need to be
specified during the queue creation and not after.

Merging these endpoints together is enough to fix this issue.

DocImpact

Implements blueprint: api-v1.1-merge-queue-and-metadata

Change-Id: Icca1a2b2066a518f8f50cad20e55b39a8bff0f2d