Enable MongoDB tests on py27
This patch sets the ZAQAR_TEST_MONGODB env variable just for
the py27 test. We would enable it for all envs, but only the
py27 gate has a recent enough version of MongoDB available.

We considered automatically detecting the mongo version in
code and skipping tests if the version was incompatible, but
that proved to be problematic since stevedore masks
exceptions that get raised by a driver class' __init__
method.

Change-Id: Ib751048a9ed1d7bb33ef80a5ed9e9cf7bdf71d57