Zaqar configuration reference docs
This change adds the basic configuration reference
in the user guide.

It includes the document structure outline according
the OpenStack configuration reference standard and
a list of configuration options present in zaqar.conf
with a description of default values and value types.

Configuration option tables has been generated using
the autogenerate_config_docs tool in openstack-doc-tools
repo. Needed files to easier update this tables in the
near future are located in the autogen folder.

blueprint document-config-options

Change-Id: I17884400f711f6d2723b0774c20e7df4ffb0e812