Adds the API reference to the devs guide
This fix adds the detailed API reference to the developers guide
as it's done in major projects in the stack.

It also fix Sphinx warning and errors and adds a .rst files
generator to easily update the api reference in the future.

Plus, it includes an index for API, modules and search.

Co-Authored-By: Flavio Percoco <fpercoco@redhat.com>

Change-Id: If42121c9e8e852785cb2db3aa5d6895223b9d604