Do not fail silently when given a bogus claim ID
This patch modifies the message delete operation so that it raises
an exception when given a totally bogus claim ID. Previously, the
drivers would just fail silently. On the other hand, the drivers
would raise an error if the claim ID was formatted correctly, but
the message was no longer claimed.

In both scenarious, the client needs a hint that something went
wrong, and so this patch simply has them both result in the same
error. The WSGI transport driver was also updated to return more
accurate error responses.

       / .'
 .---. \/
(._.' \()
 ^"""^"

Change-Id: I6f64c5eb764c821f3996c0ca6789b49d84a6cd2c