Implement claim and message expiration logic for Redis
This patch uses Redis where possible to automatically expire
claim and message data. Where it is not possible, manual
GC functions have been added in order to iterate through the
remaining data and clean it up.

Note that this patch does not include an implementation for
a command script that will call the Redis drivers' GC
method. That will be done in a followup patch.

Also, the GC functions in this patch do not decrement the
stats counters, since those are actually removed in the next
patch.

This change has been tested locally on my box using the
ZAQAR_TEST_REDIS=1 flag.

Change-Id: I38bfc52cf2de0ed078478aca80e6ad56b5bd55aa
Partially-Implements: blueprint redis-storage-driver