Fix Redis message controller getting stuck in while loop
The Redis driver was getting stuck in a loop while trying to create
a claim. The root cause turned out to be some logic that was trying
to skip past all the claimed messages at the head of the queue, but
did not advance the offset if no unclaimed messages were found,
leading to a neverending loop.

Change-Id: I8d2e1a7820c30e92a63ed938013f0e69dff83b2f
Partially-Implements: blueprint redis-storage-driver