Add ZAQAR_TEST_EVERYTHING to enable all tests
We've a way to enable some tests that are environment dependant but
there's no good way, other than setting all those variables, to enable
all tests in a specific environment.

This patch adds a new variable to do the above. It also DRYes a bit the
check of env variables that have been set.

Change-Id: Ia75925b4e22e27dab4d67caf67db80c191717caf