Add a dox.yml config file
Dox is a new tool similar to tox but on docker. One of the benefits is
that it eliminates the need of installing all the package dependencies,
which removes the burden from developers.

The tool is still under heavy development but since Zaqar's test suite
is clean and organized, it may be a good project for early adoption.

Change-Id: Iec2698448618fee3019a6877c5c2ae87fda677b6