Update Zaqar backend requirements
MongoDB is not Zaqar's default storage backend.
In fact, Zaqar don't have a default storage backend.
Considering this, pymongo shouldn't be in requirements.txt.
We still need it for the tests environments though.

Documentation about this situation will be added in the developers
guide and in the users guide in a follow up patch.

Change-Id: Ia9e039c671274fa32d90d8c31be6dca4d47bcbaf