Don't register options in __init__
Each driver currently registers its own config options in the
constructor. This is not a good thing and has caused things like
`dynamic_conf` to happen.

This patch adds a method that is called in the constructor of the base
class that goes through the driver's options and registers them. The
method catches duplicated exceptions. We don't care about those for now
and we assume the driver consumer knows what it's doing.

Bug: #1372335
Change-Id: I3341416332a0e620e428308f463d1bb635927a5d