Add first reliability enforcement
This patch adds an reliability enforcement for mongodb's driver. It
forces deployers to use replicasets or mongos as a mongodb cluster for
Zaqar. In addition to that, it forces deployers to provide a write
concern > 2 and/or majority.

If none of this are met, the driver will raise a RuntimeException and
fail to start. If no write concern is provided, majority will be used.

Change-Id: Ie74a4b441654243b3ed7e7fd6c40863969cd446d
Closes-bug: #1372335
(cherry picked from commit a1163331fcb3253ae6af5e4cb8c528eb3679baf1)