Add support for redis-sentinel
This patch adds the ability to configure Sentinel support in
zaqar.conf. The code has been tested using one and two sentinels
managing one master and one slave. Failovers from master to slave
were intiated and it was verified that clients can continue to
query the service with no downtime in the case of a lost Redis
node.

Co-Authored-By: Kurt Griffiths <kurt.griffiths@rackspace.com>
Closes-Bug: 1367020
Change-Id: Idcaaab3ba17c84a31b918ad8376f1d5fc105b53a