Move marker field to the first position (index)
This patch moves the marker field `k` to the top-most position of the
`MARKER_INDEX_FIELDS` compound index. This change improves the index in
2 ways:

1. It'll start filtering for messages starting with marker `k`. This is
good because the probability of having millions of messages with the
same marker is lower than having millions of messages for the same
project. This means fewer index records will need to be scanned to get
the messages.

2. This key will be used as a shard-key because it brings enough
cardinality to ensure chunks are evenly distributed and it also ensures
uniqueness of markers per p_q

Change-Id: I13d3773f7e5a2835f4d57a0e27b5dad221234028