Fix ttl and grace story for claims updates
Claim's updates are currently extending the life of a claim by extending
its `ttl` without taking into account the claim's grace. In addition to
this, the `ttl`'s default value is 60 for claim's renewal instead of
being the value of the configured default ttl.

This patch allows users to specify a `grace` when renewing a claim. If
`grace` is not present, it'll use the configured default value for it.

The change doesn't break backwards compatibility as in v1_0 a ttl *is*
required for claims renewal.

Closes-bug: #1356184
Change-Id: I9cce93c65f6d14ed8fabbbbcb9bf7ddc3bc11af2