Test message delete when claim expires
The patch adds a test for message deletion after a claim has expired.
The result of the test is that the message should be deleted.

Unfortunately, some issues related to the state of the sqlalchemy driver
make it quite hard for this test to work there without further changes.
I'm proposing this patch with some skips in place for the scenarios
where sqlalchemy is required, I'll propose a follow up patch to fix
those issues and enable these tests.

Closes-bug: #1369536

Change-Id: Icb12eed18c37455ad3528fa6a84fb8f2d8fcf19f