Replace sqlite with sqlalchemy
We used to have an sqlite driver and we kept it as the default value for
our storage drivers to keep backwards compatibility. It is now time to
move forward and use sqlalchemy instead.

Note that the entry points in setup.cfg are kept because we need them to
load the proper driver from pools based on the connection uri scheme.

Change-Id: I527c6ab15cc36e69d49a57a916ebc1296f73e3ce