Don't register pool endpoints when pooling=False
We recently enabled admin_mode=True in devstack. This raised an issue
with the way we register endpoints. The redis driver has no support for
pooling yet and this made Zaqar fail when running in admin mode because
it tried to register pool endpoints as well.

Although this issue is a side-effect caused by the missing feature, it
does represent an error and a need to improve the way endpoints are
registered.

Change-Id: I0f65bb25383b889048c900daa657121f64234663
Closes-bug: #1399554