API handler and API v1.1 Queue endpoints
This change adds the API handler module. This module will be the
connection between the transport and the API. The transport interacts
with this module by creating a Request and calling the process_request
function. The API, then, is in charge of processing that Request.
The API expects that the data in the request is already validated by
the transport.

In an effort to make this extensible, each API version has to define
it own endpoint module. This module interacts with the storage directly.

A Response will be returned to the transport.

In the following patches the rest of the endpoints will be added and
the validation of the Requests will be implemented.

Change-Id: I48b1937d7f5dec058879960bc556dee2a9c35b70
Partial-Implements: blueprint cross-transport-api-spec