Expose healt in homedoc
Healt is missing from the v1.1 and v2 homedoc. This patch adds
it just when admin_mode is enabled.

This patch also fix a broken test.
The test_href_template test assumes that every resource has
an href-template property but in some cases, like for the
healt resource, you just need href.

Change-Id: I2cbd69a9a3d4ddac374d8fadb3b0670f3078456e
Close-bug: #1403486