Remove sqlalchemy from our data plane
This patch removes the sqlalchemy as a valid data driver. In order to
make this work properly, we also need the data plane to be properly
split into 2 completely independent pipelines.

This patch, unfortunately, disables some of the sqlalchemy's tests for
queues and pools. The reason is that, until these planes are correctly
separated and pooling is enabled by default, we won't be able to
properly run those tests. A follow-up patch, expected to land in kilo,
will do this.

DocImpact

Implements blueprint: disable-sqlalchemy-as-pool

Change-Id: I5ca73102241692b0e1b23b3eaaaf1938f0106dfa