Deprecate v1
This patch marks v1 as deprecated. It'll be removed in the next cycle
and we won't  be fixing things for it anymore. This change should be
reflected in the client library as well.

In this patch, devstack has been switched to use v2 directly, which is
the current developed/maintained version.

APIImpact

Change-Id: I2f32828835b3329b37bf7bac04024c46c161054a