Move common configs under common.configs
We've some options in bootstrap and other places that should probably go
to a common place so that we can guarantee they are always registered
when the Zaqar starts.

This patch adds a new zaqar.common.config module that holds all the
common configs but doesn't register them. This step is still performed
when it is required.

Change-Id: I9941095ed56835093d56b5ad2bf85b53bdda0c8b
Partially-Implements: pre-signed-url