Fix enabling of Zaqar in Devstack plugin
Zaqar can be installed in two ways using devstack:
- using old approach called "hooks"
- using new approach called "devstack plugin"

So, if we try to install it using only devstack plugin enabling it as following:

"enable_plugin zaqar https://github.com/openstack/zaqar"

in localrc file, then we get nothing in a result. Because Devstack plugin tries
to enable zaqar using 'enable_service zaqar' [1].
But, then, devstack performs action verifying that 'zaqar-server' is
enabled. [2]
This leads to that zaqar is not installed at all.

For the moment, Zaqar installation works in CI because 'zaqar-server' is
enabled by other means using hooks.

[1]: https://github.com/openstack/zaqar/blob/bfcc60c1/devstack/settings#L4
[2]: https://github.com/openstack/zaqar/blob/bfcc60c1/devstack/plugin.sh#L246

Change-Id: I3c55756171e772d5c9e69066e0434ace5da0f852
Closes-Bug: #1476583