Move _(get|inc)_counter out of QueueController
These 2 counter utilities are used in the message controller as a way to
paginate and guarantee FIFO. The counter was kept in the queue record
since it provides a centralized place for it to live in. However, since
we moved the QueueController to the control plane, we can't expect the
QueueController to be on the same sotrage as the MessageController.

The patch keeps the old methods around for backwards compatibility.
It'll check if the QueueController has these methods and call those
rather than the new ones to avoid reseting the counter. This should
eventually go away but for now, it's probably the best we can do.

Closes-bug: #1484826
Change-Id: I57a9c4baa1d6c98fb938526ed90d89b6e0e21ed9