Don't generate docs for zaqar.transport.wsgi.app
We've been generating docs for the `zaqar.transport.wsgi.app` module,
which tries to create an instance of the wsgi transport, which then
tries to create an instance of the data controller and connect to the
database.

This module does not contain any docs and we shouldn't require a running
DB instance to generate these docs.

Change-Id: Id1c3cabc61a790bc0e5cccb22e660c933788ded8
Closes-bug: #1489426