Cleanup of Translations
In preparation for the release, do some cleanups for translations.

Removes all po files that are partially translated. The translation
team has decided to exclude files with less than 66 % of translated
content. There is no content lost, all data is in the translation
server, we just remove it from the repository.

This updates also recreates all pot (translation source files) to
reflect the state of the repository.

This change needs to be done manually since the automatic import does
not handle some of these cases.

Depends-On: Icf355f9a6d667371a4f2b6077ca4954e1351e92d
Change-Id: I6f6662af780a002ed37c79bdd8fdd12fac110d4b