Make websocket run in correct logging mode
Currently websocket always runs in debug mode.
It is because the method variable is passed as a "debug" parameter to
websocket constructor instead of boolean variable.
So websocket will interpret this parameter as true and debug mode will
be on.

This commit makes sure the variable passed as "debug" parameter to
websocket constructor is boolean and really tells if zaqar-server's
debug logging is enabled.
This boolean variable is aquired from the information in local logger
instance.

Change-Id: I701f986a85bdb024c3757ede5a388a78791343aa
Closes-Bug: 1505263