Fix tox not performing some tests in some Envs
Many tests are being skipped while testing using tox in some
environments.
That's because some important environment variables are not defined for
these environments:
pypy, py34.
The variables are:
ZAQAR_TEST_MONGODB=1
ZAQAR_TEST_SLOW=1
This change defines these environment variables in tox.ini. And not only
for the few environments that need such variables, but for all
environments. Because it's the most simple way to do it in which I see
no harm.
Also this change deletes deprecated python environments (py26 and py33)
from the list of environments.

Depends-On: I37937ffe9b71323a92880f678e84c7c24e6872d6
Change-Id: I3a8cef038ed9a37c8dc29850dfdda31b72c05883
Closes-Bug: 1508519