Add `_max_messages_post_size` and `_default_message_ttl` for queue
Now Zaqar supports metadata as the attributes of a queue, so that user
can fill it with some key-value pair. However, Zaqar itself isn't
using it very much(only the '_flavor'). Meanwhile, we have some
configurations defined in zaqar.conf against messages or claims which
are global defined. For example, the max message size. So the idea is
adding more reserved attributes for queue so that the queues of Zaqar
are more flexible. This patch adds the `_max_messages_post_size` and
`_default_message_ttl` for message posting, tests are added as well.

APIImpact
DocImpact

blueprint more-reserved-queue-attributes

Change-Id: I7ba84b6e862ff15684a1b1d919c0fed60f4c58b2