Ensure JSON sent over websocket is a dictionary
The json lib in python can transform the input into many kinds of
data struct, such as int,str,boolean and so on. But only dict can
be handled by websocket. Now if the input is not dict like, the
zaqar-server will be stop at all. So before handle the request,
we should ensure the input format is acceptable.

Closes-bug: #1537965
Change-Id: I4e8bf4c0477741b40df3c58abb32ba73c792256c