Support metadata update issue for v2
Now there is no way to update metadata after queue create.
That because we were trying to remove metadata but finally we
decided to keep it. As a result, we missed a way to update queue's
metadata. This patch fixes it buy supporting PATCH for queue. And
a test case is added.

APIImpact
DocImpact

Closes-Bug: #1543900

Change-Id: Iba6d7a0ca5cf7a755c20d9af69d5c52687142dff