Show 'age' field in subscriptions (redis)
We recently fixed subscriptions and now they are expirable resources as
they should be.

This patch adds 'age' field to subscription resource representation
which is exposed to the user and defined in redis driver.

This patch also removes existing 'expires' field, because:
1. The 'expires' field is missing in mongodb's subscription resource
representation. (But age field was also missing until a recent patch).
2. The 'expires' field has int (timestamp) format, which is not human
readable.
3. Other Zaqar's expirable resources show 'age' and 'ttl' fields instead
of showing 'expires' field.
4. There is no need in 'expires' field when 'age' and 'ttl' fields are
exposed to the user.

This change in unlikely to affect backward compatibility, because
subscription expirations started to work only recently.

Closes-Bug: 1552845
APIImpact
DocImpact
Change-Id: I18f908b2f137bd4b8e5b3d3275063a1d37942093