Fix PATCH queue's metadata
Currently, the way queue's PATCH HTTP method is implemented in API v2 is
not correct.

This patch makes PATCH HTTP method to work as RFC6902 recommends:
http://tools.ietf.org/html/rfc6902
This patch adds tests.
This patch makes Zaqar return 404 response in the case when the user
tries to perform PATCH on non-existing queue.

APIImpact
DocImpact

Closes-Bug: #1553002
Co-Authored-By: Fei Long Wang <flwang@catalyst.net.nz>
Co-Authored-By: Eva Balycheva <ubershy@gmail.com>
Change-Id: Id1f02c63e8f078ddb657d3e9bc83ac0668672748