Add reserved metadata check
Now,Zaqar will always return "max_messages_post_size"
and "default_message_ttl" when get a queue's metadata.
So users can't know whether these values are in db or
not. Then if they take the "replace" or "remove" action
to update the queue's metadata when these two values are
not in db, they will get an error: "can not update(remove)
non-exist metadata".

So we should check them before the queue update action.

Change-Id: Ia97e80a406c8321c7e8d0aa2652023049e5ee4c0
Closes-bug: #1570160