Reword description of grace
Now Zaqar uses "claim expiration time > current message TTL"
to judge whether to extend message ttl or not.
The explanation of grace "If a claimed message would normally
live longer than the grace period, its expiration is not adjusted."
is not very clear to users.

So we reword the description of grace to "If a claimed message
would normally live longer than the claim's live period,
its expiration is not adjusted."

Change-Id: Ib9b042e99116fb4c4656696d413a778e4f07f4c6
Closes-Bug: #1632629