swift driver: rely on keystone session for swift connection
Even though the swiftclient can take a token, it doesn't handle
reauthentication properly, and neither does the code in the swift
driver. This is done by the keystone session object, and we should leave
the reauthentication handling there.

We then pass instead the session object to the swift client
connection, avoiding reauthentication issues.

The session is initialized lazily only when the driver is attempted to
be used.

Change-Id: Ie767be7fbe311cfbb9e73fe18641577626880162