Simplify SQL run() method
Currently, Zaqar is using only Core SQL in an autocommit
context, meaning, no ORM, and when a statement is run,
it's on a new pool checkout, autocommit, returned.

In that case there's no reason to deal with the ORM
Session or thread-local scope of scoped_session() since there
is no state being used across requests here, other than the
"engine" itself.

Keeping in mind Zaqar wants to move to a transactionalized
pattern w/ context managers, for now they are doing the equivalent
of connectionless execution.   Just switch to that for now.

Change-Id: I06bfb96381864448a874278364f06489bdc76ddd