Fix claims of MongoDB when enabling DLQ
When the dead letter queue enabled and part of messages meet their
max claim count, these messages will be moved to dead letter queue
and current claim request will remove None for the claim ID and
empty list for the messages based on current implement. But it's
not correct if there are messages have been claimed successfully.
This patch fixes the issue and changes the test code accordingly.

Closes-Bug: 1707106

Change-Id: I34c207755f267b37cbab080aefe9688680409dcb