Redis connection support password configure in zaqar
Redis connection doesn't support password configure in zaqar,
so redis-server can not set a password. If redis service doesn't
set a password, it will suffer a large number of attacks.
The patch will support password configure in zaqar. In addition,
it is needed to add unit test in the previous
patch[https://review.openstack.org/#/c/511676/].

Change-Id: I03661584f1c24fdb3e76a819c9eaa0ed32f272c4