Add .stestr.conf configuration
os-testr has moved over to use stestr instead of testr. While this
is usually compaible with existing settings, there is a warning
that is emitted when .stestr.conf is not present. It is usually
able to fall back to parsing the .testr.conf file, but to be more
correct and to prevent future problems we should update the config.

Change-Id: I95a6d1135fb4861bd3fb2086ae7a63e486130db0
Closes-Bug: #1809062