Unblock gate failures
There are a few issues causing all gate runs to fail that are addressed
in this patch.

Update constraints to match global requirements:

There are a few local requirements that conflict with what is now in
upstream global-requirements. This updates sphinx to match the
constrained versions due to incompatibility with py2 and uncaps
jsonschema.

Update handling for falcon 2.0.0 behavior changes:

The 2.0.0 release of falcon changed the default values for two request
query parsing options that would cause failures due to unexpected
results. This sets those options to their previous defaults to maintain
compatibility across 2.0.0 and pre-2.0.0 versions of the package.
Lower constraint version raised to 1.1.0 where the options were
introduced.

There are additional integration test failures with Falcon 2.0.0 that
were due to changes in the subscription URL path.

Closes-bug: #1830903

Requirements and upper constraints not enforced in some tox targets:

The docs and api-ref tox targets were not setting the right dependencies
or the upper constraints to use for pip installs. Also updates the
constraints to the newer preferred static redirect URL maintained by the
release team.

Co-authored-by: wanghao <sxmatch1986@gmail.com>
Change-Id: I19c5363878c629a04e1be902c865ee7cea92dcd9
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>