Native Zuul v3 version of tempest and rally jobs
Use the native Zuul v3 jobs defined in the zaqar-tempest-plugin
repository, so that they can be used on all branches.

Use rally-task-zaqar provided by rally-openstack
as base for zaqar-rally.

Clean up the legacy files, but leave the hooks which are also
used by the grenade jobs. They will be cleaned separately.

Depends-On: https://review.opendev.org/747543
Change-Id: I31dae00091b0de7bf723c086b1c81d5500601dd3